+++ app/src/main/java/CS2263/CourseProject/Subtask.java
@@ -1,0 +1,11 @@
+package CS2263.CourseProject;
+
+import lombok.Getter;
+import lombok.Setter;
+
+
+public class Subtask extends Task {
+    @Getter @Setter private Task parentTask;
+
+
+}