+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 Running log of all changes made to the CS2263 Course Project.
 
 ## Unreleased
-* // UNRELEASED FEATURES GO HERE
+*  Task and Subclass classses added
+* Lombok dependencies added
+* 3/31-- Task, Subtask, and Login Credentials Completed
 
-## // VERSION NUM - // DATE
+## // 1.0.0
 ### Fixed
 * // FIXES TO CONTENT ALREADY IN THE PROGRAM GOES HERE